Reverting some files that were accidentally pushed
This reverts some of commit ef5beaa746a962c3c04c25d5243e7a7a80a69e79.

+++ neo/renderer/Material.cpp
@@ -505,9 +505,9 @@ idMaterial::ParseTerm
 Returns a register index
 =================
 */
-size_t idMaterial::ParseTerm( idLexer &src ) {
+int idMaterial::ParseTerm( idLexer &src ) {
 	idToken token;
-	size_t		a, b;
+	int		a, b;
 
 	src.ReadToken( &token );
