+++ neo/ui/Window.cpp
@@ -1715,38 +1715,38 @@ idWindow::GetWinVarOffset
 ================
 */
 int idWindow::GetWinVarOffset( idWinVar *wv, drawWin_t* owner) {
-	size_t ret = -1;
+	int ret = -1;
 
 	if ( wv == &rect ) {
-		ret = (size_t)&( ( idWindow * ) 0 )->rect;
+		ret = (int)&( ( idWindow * ) 0 )->rect;
 	}
 
 	if ( wv == &backColor ) {
-		ret = (size_t)&( ( idWindow * ) 0 )->backColor;
+		ret = (int)&( ( idWindow * ) 0 )->backColor;
 	}
 
 	if ( wv == &matColor ) {
-		ret = (size_t)&( ( idWindow * ) 0 )->matColor;
+		ret = (int)&( ( idWindow * ) 0 )->matColor;
 	}
 
 	if ( wv == &foreColor ) {
-		ret = (size_t)&( ( idWindow * ) 0 )->foreColor;
+		ret = (int)&( ( idWindow * ) 0 )->foreColor;
 	}
 
 	if ( wv == &hoverColor ) {
-		ret = (size_t)&( ( idWindow * ) 0 )->hoverColor;
+		ret = (int)&( ( idWindow * ) 0 )->hoverColor;
 	}
 
 	if ( wv == &borderColor ) {
-		ret = (size_t)&( ( idWindow * ) 0 )->borderColor;
+		ret = (int)&( ( idWindow * ) 0 )->borderColor;
 	}
 
 	if ( wv == &textScale ) {
-		ret = (size_t)&( ( idWindow * ) 0 )->textScale;
+		ret = (int)&( ( idWindow * ) 0 )->textScale;
 	}
 
 	if ( wv == &rotate ) {
-		ret = (size_t)&( ( idWindow * ) 0 )->rotate;
+		ret = (int)&( ( idWindow * ) 0 )->rotate;
 	}
 
 	if ( ret != -1 ) {
@@ -2827,7 +2827,7 @@ idWindow::ParseTerm
 Returns a register index
 =================
 */
-size_t idWindow::ParseTerm( idParser *src,	idWinVar *var, int component ) {
+int idWindow::ParseTerm( idParser *src,	idWinVar *var, int component ) {
 	idToken token;
 	int		a, b;
 
@@ -2872,7 +2872,7 @@ size_t idWindow::ParseTerm( idParser *src,	idWinVar *var, int component ) {
 		var = GetWinVarByName(token, true);
 	}
 	if (var) {
-		a = (size_t)var;
+		a = (int)var;
 		//assert(dynamic_cast<idWinVec4*>(var));
 		var->Init(token, this);
 		b = component;
@@ -2902,7 +2902,7 @@ size_t idWindow::ParseTerm( idParser *src,	idWinVar *var, int component ) {
 		// ugly but used for post parsing to fixup named vars
 		char *p = new char[token.Length()+1];
 		strcpy(p, token);
-		a = (size_t)p;
+		a = (int)p;
 		b = -2;
 		return EmitOp(a, b, WOP_TYPE_VAR);
 	}
@@ -3776,35 +3776,35 @@ void idWindow::FixupTransitions() {
 		transitions[i].data = NULL;
 		if ( dw && ( dw->win || dw->simp ) ){
 			if ( dw->win ) {
-				if ( transitions[i].offset == (size_t)&( ( idWindow * ) 0 )->rect ) {
+				if ( transitions[i].offset == (int)&( ( idWindow * ) 0 )->rect ) {
 					transitions[i].data = &dw->win->rect;
-				} else if ( transitions[i].offset == (size_t)&( ( idWindow * ) 0 )->backColor ) {
+				} else if ( transitions[i].offset == (int)&( ( idWindow * ) 0 )->backColor ) {
 					transitions[i].data = &dw->win->backColor;
-				} else if ( transitions[i].offset == (size_t)&( ( idWindow * ) 0 )->matColor ) {
+				} else if ( transitions[i].offset == (int)&( ( idWindow * ) 0 )->matColor ) {
 					transitions[i].data = &dw->win->matColor;
-				} else if ( transitions[i].offset == (size_t)&( ( idWindow * ) 0 )->foreColor ) {
+				} else if ( transitions[i].offset == (int)&( ( idWindow * ) 0 )->foreColor ) {
 					transitions[i].data = &dw->win->foreColor;
-				} else if ( transitions[i].offset == (size_t)&( ( idWindow * ) 0 )->borderColor ) {
+				} else if ( transitions[i].offset == (int)&( ( idWindow * ) 0 )->borderColor ) {
 					transitions[i].data = &dw->win->borderColor;
-				} else if ( transitions[i].offset == (size_t)&( ( idWindow * ) 0 )->textScale ) {
+				} else if ( transitions[i].offset == (int)&( ( idWindow * ) 0 )->textScale ) {
 					transitions[i].data = &dw->win->textScale;
-				} else if ( transitions[i].offset == (size_t)&( ( idWindow * ) 0 )->rotate ) {
+				} else if ( transitions[i].offset == (int)&( ( idWindow * ) 0 )->rotate ) {
 					transitions[i].data = &dw->win->rotate;
 				}
 			} else {
-				if ( transitions[i].offset == (size_t)&( ( idSimpleWindow * ) 0 )->rect ) {
+				if ( transitions[i].offset == (int)&( ( idSimpleWindow * ) 0 )->rect ) {
 					transitions[i].data = &dw->simp->rect;
-				} else if ( transitions[i].offset == (size_t)&( ( idSimpleWindow * ) 0 )->backColor ) {
+				} else if ( transitions[i].offset == (int)&( ( idSimpleWindow * ) 0 )->backColor ) {
 					transitions[i].data = &dw->simp->backColor;
-				} else if ( transitions[i].offset == (size_t)&( ( idSimpleWindow * ) 0 )->matColor ) {
+				} else if ( transitions[i].offset == (int)&( ( idSimpleWindow * ) 0 )->matColor ) {
 					transitions[i].data = &dw->simp->matColor;
-				} else if ( transitions[i].offset == (size_t)&( ( idSimpleWindow * ) 0 )->foreColor ) {
+				} else if ( transitions[i].offset == (int)&( ( idSimpleWindow * ) 0 )->foreColor ) {
 					transitions[i].data = &dw->simp->foreColor;
-				} else if ( transitions[i].offset == (size_t)&( ( idSimpleWindow * ) 0 )->borderColor ) {
+				} else if ( transitions[i].offset == (int)&( ( idSimpleWindow * ) 0 )->borderColor ) {
 					transitions[i].data = &dw->simp->borderColor;
-				} else if ( transitions[i].offset == (size_t)&( ( idSimpleWindow * ) 0 )->textScale ) {
+				} else if ( transitions[i].offset == (int)&( ( idSimpleWindow * ) 0 )->textScale ) {
 					transitions[i].data = &dw->simp->textScale;
-				} else if ( transitions[i].offset == (size_t)&( ( idSimpleWindow * ) 0 )->rotate ) {
+				} else if ( transitions[i].offset == (int)&( ( idSimpleWindow * ) 0 )->rotate ) {
 					transitions[i].data = &dw->simp->rotate;
 				}
 			}
@@ -3864,7 +3864,7 @@ void idWindow::FixupParms() {
 			const char *p = (const char*)(ops[i].a);
 			idWinVar *var = GetWinVarByName(p, true);
 			delete []p;
-			ops[i].a = (size_t)var;
+			ops[i].a = (int)var;
 			ops[i].b = -1;
 		}
 	}
